Rename og lagre form values
