+++ src/index.less
@@ -22,7 +22,7 @@
 @import "pages/forside/sections/3-header/Header.less";
 @import "pages/forside/sections/4-personinfo/3-adresser/Adresser.less";
 @import "pages/forside/sections/4-personinfo/1-personalia/Personalia.less";
-@import "pages/forside/sections/4-personinfo/2-kontaktinfo/subsections/kontakt-og-reservasjonsregisteret/KontaktInfo.less";
+@import "pages/forside/sections/4-personinfo/2-kontaktinfo/subsections/kontakt-og-reservasjonsregisteret/DKIF.less";
 @import "pages/forside/sections/4-personinfo/4-ditt-nav-kontor/DittNavKontor.less";
 @import "pages/forside/sections/5-arbeidsforhold/Arbeidsforhold.less";
 @import "pages/forside/sections/6-eksterne/EksterneLenker.less";
